% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{create_device_pool}
\alias{create_device_pool}
\title{Creates a device pool}
\usage{
create_device_pool(projectArn, name, rules, description = NULL)
}
\arguments{
\item{projectArn}{The ARN of the project for the device pool.}

\item{name}{The device pool's name.}

\item{rules}{The device pool's rules.}

\item{description}{The device pool's description.}
}
\description{
Creates a device pool.
}
\examples{
# The following example creates a new device pool named MyDevicePool
# inside an existing project.
create_device_pool(
  name = "MyDevicePool",
  description = "My Android devices",
  projectArn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  rules = list()
)

}
