% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{create_project}
\alias{create_project}
\title{Creates a new project}
\usage{
create_project(name, defaultJobTimeoutMinutes = NULL)
}
\arguments{
\item{name}{[required] The project's name.}

\item{defaultJobTimeoutMinutes}{Sets the execution timeout value (in minutes) for a project. All test runs in this project will use the specified execution timeout value unless overridden when scheduling a run.}
}
\description{
Creates a new project.
}
\section{Accepted Parameters}{
\preformatted{create_project(
  name = "string",
  defaultJobTimeoutMinutes = 123
)
}
}

\examples{
# The following example creates a new project named MyProject.
\donttest{create_project(
  name = "MyProject"
)}

}
