% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{delete_device_pool}
\alias{delete_device_pool}
\title{Deletes a device pool given the pool ARN}
\usage{
delete_device_pool(
  arn = "string"
)
}
\arguments{
\item{arn}{[required] Represents the Amazon Resource Name (ARN) of the Device Farm device pool you wish to delete.}
}
\description{
Deletes a device pool given the pool ARN. Does not allow deletion of curated pools owned by the system.
}
\examples{
# The following example deletes a specific device pool.
delete_device_pool(
  arn = "arn:aws:devicefarm:us-west-2::devicepool:123-456-EXAMPLE-GUID"
)

}
