% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{delete_instance_profile}
\alias{delete_instance_profile}
\title{Deletes a profile that can be applied to one or more private device instances}
\usage{
delete_instance_profile(
  arn = "string"
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the instance profile you are requesting to delete.}
}
\description{
Deletes a profile that can be applied to one or more private device instances.
}
\examples{

}
