% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{delete_remote_access_session}
\alias{delete_remote_access_session}
\title{Deletes a completed remote access session and its results}
\usage{
delete_remote_access_session(
  arn = "string"
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the sesssion for which you want to delete remote access.}
}
\description{
Deletes a completed remote access session and its results.
}
\examples{
# The following example deletes a specific remote access session.
delete_remote_access_session(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:session:EXAMPLE-GUID-123-456"
)

}
