% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{delete_run}
\alias{delete_run}
\title{Deletes the run, given the run ARN}
\usage{
delete_run(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) for the run you wish to delete.}
}
\description{
Deletes the run, given the run ARN.
}
\details{
\strong{Note} Deleting this resource does not stop an in-progress run.
}
\section{Accepted Parameters}{
\preformatted{delete_run(
  arn = "string"
)
}
}

\examples{
# The following example deletes a specific test run.
\donttest{delete_run(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:run:EXAMPLE-GUID-123-456"
)}

}
