% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{delete_upload}
\alias{delete_upload}
\title{Deletes an upload given the upload ARN}
\usage{
delete_upload(arn)
}
\arguments{
\item{arn}{Represents the Amazon Resource Name (ARN) of the Device Farm upload you wish to delete.}
}
\description{
Deletes an upload given the upload ARN.
}
\examples{
# The following example deletes a specific upload.
delete_upload(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:upload:EXAMPLE-GUID-123-456"
)

}
