% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{delete_vpce_configuration}
\alias{delete_vpce_configuration}
\title{Deletes a configuration for your Amazon Virtual Private Cloud (VPC) endpoint}
\usage{
delete_vpce_configuration(
  arn = "string"
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the VPC endpoint configuration you want to delete.}
}
\description{
Deletes a configuration for your Amazon Virtual Private Cloud (VPC) endpoint.
}
\examples{

}
