% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_device}
\alias{get_device}
\title{Gets information about a unique device type}
\usage{
get_device(arn)
}
\arguments{
\item{arn}{[required] The device type's ARN.}
}
\description{
Gets information about a unique device type.
}
\section{Accepted Parameters}{
\preformatted{get_device(
  arn = "string"
)
}
}

\examples{
# The following example returns information about a specific device.
\donttest{get_device(
  arn = "arn:aws:devicefarm:us-west-2::device:123EXAMPLE"
)}

}
