% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_device_instance}
\alias{get_device_instance}
\title{Returns information about a device instance belonging to a private device fleet}
\usage{
get_device_instance(arn)
}
\arguments{
\item{arn}{The Amazon Resource Name (ARN) of the instance you're requesting information about.}
}
\description{
Returns information about a device instance belonging to a private device fleet.
}
\examples{

}
