% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_instance_profile}
\alias{get_instance_profile}
\title{Returns information about the specified instance profile}
\usage{
get_instance_profile(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of your instance profile.}
}
\description{
Returns information about the specified instance profile.
}
\section{Accepted Parameters}{
\preformatted{get_instance_profile(
  arn = "string"
)
}
}

