% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_job}
\alias{get_job}
\title{Gets information about a job}
\usage{
get_job(arn)
}
\arguments{
\item{arn}{[required] The job's ARN.}
}
\description{
Gets information about a job.
}
\section{Accepted Parameters}{
\preformatted{get_job(
  arn = "string"
)
}
}

\examples{
# The following example returns information about a specific job.
\donttest{get_job(
  arn = "arn:aws:devicefarm:us-west-2::job:123-456-EXAMPLE-GUID"
)}

}
