% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_network_profile}
\alias{get_network_profile}
\title{Returns information about a network profile}
\usage{
get_network_profile(arn)
}
\arguments{
\item{arn}{The Amazon Resource Name (ARN) of the network profile you want to return information about.}
}
\description{
Returns information about a network profile.
}
\examples{

}
