% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_run}
\alias{get_run}
\title{Gets information about a run}
\usage{
get_run(
  arn = "string"
)
}
\arguments{
\item{arn}{[required] The run's ARN.}
}
\description{
Gets information about a run.
}
\examples{
# The following example gets information about a specific test run.
get_run(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:run:5e01a8c7-c861-4c0a-b1d5-5ec6e6c6dd23/0fcac17b-6122-44d7-ae5a-12345EXAMPLE"
)

}
