% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_suite}
\alias{get_suite}
\title{Gets information about a suite}
\usage{
get_suite(
  arn = "string"
)
}
\arguments{
\item{arn}{[required] The suite's ARN.}
}
\description{
Gets information about a suite.
}
\examples{
# The following example gets information about a specific test suite.
get_suite(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:suite:EXAMPLE-GUID-123-456"
)

}
