% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_upload}
\alias{get_upload}
\title{Gets information about an upload}
\usage{
get_upload(arn)
}
\arguments{
\item{arn}{The upload's ARN.}
}
\description{
Gets information about an upload.
}
\examples{
# The following example gets information about a specific upload.
get_upload(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:upload:EXAMPLE-GUID-123-456"
)

}
