% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_vpce_configuration}
\alias{get_vpce_configuration}
\title{Returns information about the configuration settings for your Amazon Virtual Private Cloud (VPC) endpoint}
\usage{
get_vpce_configuration(
  arn = "string"
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the VPC endpoint configuration you want to describe.}
}
\description{
Returns information about the configuration settings for your Amazon Virtual Private Cloud (VPC) endpoint.
}
\examples{

}
