% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_instance_profiles}
\alias{list_instance_profiles}
\title{Returns information about all the instance profiles in an AWS account}
\usage{
list_instance_profiles(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{An integer specifying the maximum number of items you want to return in the API response.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Returns information about all the instance profiles in an AWS account.
}
\section{Accepted Parameters}{
\preformatted{list_instance_profiles(
  maxResults = 123,
  nextToken = "string"
)
}
}

