% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_jobs}
\alias{list_jobs}
\title{Gets information about jobs for a given test run}
\usage{
list_jobs(arn, nextToken = NULL)
}
\arguments{
\item{arn}{The run's Amazon Resource Name (ARN).}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Gets information about jobs for a given test run.
}
\examples{
# The following example returns information about jobs in a specific
# project.
list_jobs(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)

}
