% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_network_profiles}
\alias{list_network_profiles}
\title{Returns the list of available network profiles}
\usage{
list_network_profiles(
  arn = "string",
  type = "CURATED"|"PRIVATE",
  nextToken = "string"
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the project for which you want to list network profiles.}

\item{type}{The type of network profile you wish to return information about. Valid values are listed below.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Returns the list of available network profiles.
}
\examples{

}
