% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_projects}
\alias{list_projects}
\title{Gets information about projects}
\usage{
list_projects(arn = NULL, nextToken = NULL)
}
\arguments{
\item{arn}{Optional. If no Amazon Resource Name (ARN) is specified, then AWS Device Farm returns a list of all projects for the AWS account. You can also specify a project ARN.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Gets information about projects.
}
\examples{
# The following example returns information about the specified project in
# Device Farm.
list_projects(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:7ad300ed-8183-41a7-bf94-12345EXAMPLE",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)

}
