% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_runs}
\alias{list_runs}
\title{Gets information about runs, given an AWS Device Farm project ARN}
\usage{
list_runs(
  arn = "string",
  nextToken = "string"
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the project for which you want to list runs.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Gets information about runs, given an AWS Device Farm project ARN.
}
\examples{
# The following example returns information about a specific test run.
list_runs(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:run:5e01a8c7-c861-4c0a-b1d5-5ec6e6c6dd23/0fcac17b-6122-44d7-ae5a-12345EXAMPLE",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)

}
