% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_suites}
\alias{list_suites}
\title{Gets information about test suites for a given job}
\usage{
list_suites(arn, nextToken = NULL)
}
\arguments{
\item{arn}{The job's Amazon Resource Name (ARN).}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Gets information about test suites for a given job.
}
\examples{
# The following example returns information about suites, given a specific
# Device Farm project.
list_suites(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)

}
