% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_tests}
\alias{list_tests}
\title{Gets information about tests in a given test suite}
\usage{
list_tests(arn, nextToken = NULL)
}
\arguments{
\item{arn}{[required] The test suite's Amazon Resource Name (ARN).}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Gets information about tests in a given test suite.
}
\section{Accepted Parameters}{
\preformatted{list_tests(
  arn = "string",
  nextToken = "string"
)
}
}

\examples{
# The following example returns information about tests, given a specific
# Device Farm project.
\donttest{list_tests(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)}

}
