% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_uploads}
\alias{list_uploads}
\title{Gets information about uploads, given an AWS Device Farm project ARN}
\usage{
list_uploads(
  arn = "string",
  type = "ANDROID_APP"|"IOS_APP"|"WEB_APP"|"EXTERNAL_DATA"|"APPIUM_JAVA_JUNIT_TEST_PACKAGE"|"APPIUM_JAVA_TESTNG_TEST_PACKAGE"|"APPIUM_PYTHON_TEST_PACKAGE"|"APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE"|"APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE"|"APPIUM_WEB_PYTHON_TEST_PACKAGE"|"CALABASH_TEST_PACKAGE"|"INSTRUMENTATION_TEST_PACKAGE"|"UIAUTOMATION_TEST_PACKAGE"|"UIAUTOMATOR_TEST_PACKAGE"|"XCTEST_TEST_PACKAGE"|"XCTEST_UI_TEST_PACKAGE"|"APPIUM_JAVA_JUNIT_TEST_SPEC"|"APPIUM_JAVA_TESTNG_TEST_SPEC"|"APPIUM_PYTHON_TEST_SPEC"|"APPIUM_WEB_JAVA_JUNIT_TEST_SPEC"|"APPIUM_WEB_JAVA_TESTNG_TEST_SPEC"|"APPIUM_WEB_PYTHON_TEST_SPEC"|"INSTRUMENTATION_TEST_SPEC"|"XCTEST_UI_TEST_SPEC",
  nextToken = "string"
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the project for which you want to list uploads.}

\item{type}{The type of upload.

Must be one of the following values:
\itemize{
\item ANDROID\_APP: An Android upload.
\item IOS\_APP: An iOS upload.
\item WEB\_APP: A web appliction upload.
\item EXTERNAL\_DATA: An external data upload.
\item APPIUM\_JAVA\_JUNIT\_TEST\_PACKAGE: An Appium Java JUnit test package upload.
\item APPIUM\_JAVA\_TESTNG\_TEST\_PACKAGE: An Appium Java TestNG test package upload.
\item APPIUM\_PYTHON\_TEST\_PACKAGE: An Appium Python test package upload.
\item APPIUM\_WEB\_JAVA\_JUNIT\_TEST\_PACKAGE: An Appium Java JUnit test package upload.
\item APPIUM\_WEB\_JAVA\_TESTNG\_TEST\_PACKAGE: An Appium Java TestNG test package upload.
\item APPIUM\_WEB\_PYTHON\_TEST\_PACKAGE: An Appium Python test package upload.
\item CALABASH\_TEST\_PACKAGE: A Calabash test package upload.
\item INSTRUMENTATION\_TEST\_PACKAGE: An instrumentation upload.
\item UIAUTOMATION\_TEST\_PACKAGE: A uiautomation test package upload.
\item UIAUTOMATOR\_TEST\_PACKAGE: A uiautomator test package upload.
\item XCTEST\_TEST\_PACKAGE: An XCode test package upload.
\item XCTEST\_UI\_TEST\_PACKAGE: An XCode UI test package upload.
\item APPIUM\_JAVA\_JUNIT\_TEST\_SPEC: An Appium Java JUnit test spec upload.
\item APPIUM\_JAVA\_TESTNG\_TEST\_SPEC: An Appium Java TestNG test spec upload.
\item APPIUM\_PYTHON\_TEST\_SPEC: An Appium Python test spec upload.
\item APPIUM\_WEB\_JAVA\_JUNIT\_TEST\_SPEC: An Appium Java JUnit test spec upload.
\item APPIUM\_WEB\_JAVA\_TESTNG\_TEST\_SPEC: An Appium Java TestNG test spec upload.
\item APPIUM\_WEB\_PYTHON\_TEST\_SPEC: An Appium Python test spec upload.
\item INSTRUMENTATION\_TEST\_SPEC: An instrumentation test spec upload.
\item XCTEST\_UI\_TEST\_SPEC: An XCode UI test spec upload.
}}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Gets information about uploads, given an AWS Device Farm project ARN.
}
\examples{
# The following example returns information about uploads, given a
# specific Device Farm project.
list_uploads(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  nextToken = "RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"
)

}
