% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_vpce_configurations}
\alias{list_vpce_configurations}
\title{Returns information about all Amazon Virtual Private Cloud (VPC) endpoint configurations in the AWS account}
\usage{
list_vpce_configurations(
  maxResults = 123,
  nextToken = "string"
)
}
\arguments{
\item{maxResults}{An integer specifying the maximum number of items you want to return in the API response.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Returns information about all Amazon Virtual Private Cloud (VPC) endpoint configurations in the AWS account.
}
\examples{

}
