% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_package.R
\docType{package}
\name{paws.devicefarm-package}
\alias{paws.devicefarm}
\alias{paws.devicefarm-package}
\title{paws.devicefarm: AWS Device Farm}
\description{
AWS Device Farm is a service that enables mobile
    app developers to test Android, iOS, and Fire OS apps on physical
    phones, tablets, and other devices in the cloud.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
