% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{purchase_offering}
\alias{purchase_offering}
\title{Immediately purchases offerings for an AWS account}
\usage{
purchase_offering(offeringId = NULL, quantity = NULL,
  offeringPromotionId = NULL)
}
\arguments{
\item{offeringId}{The ID of the offering.}

\item{quantity}{The number of device slots you wish to purchase in an offering request.}

\item{offeringPromotionId}{The ID of the offering promotion to be applied to the purchase.}
}
\description{
Immediately purchases offerings for an AWS account. Offerings renew with the latest total purchased quantity for an offering, unless the renewal was overridden. The API returns a \code{NotEligible} error if the user is not permitted to invoke the operation. Please contact \href{mailto:aws-devicefarm-support@amazon.com}{aws-devicefarm-support@amazon.com} if you believe that you should be able to invoke this operation.
}
\examples{
# The following example purchases a specific device slot offering.
purchase_offering(
  offeringId = "D68B3C05-1BA6-4360-BC69-12345EXAMPLE",
  quantity = 1L
)

}
