% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{renew_offering}
\alias{renew_offering}
\title{Explicitly sets the quantity of devices to renew for an offering, starting from the \code{effectiveDate} of the next period}
\usage{
renew_offering(offeringId = NULL, quantity = NULL)
}
\arguments{
\item{offeringId}{The ID of a request to renew an offering.}

\item{quantity}{The quantity requested in an offering renewal.}
}
\description{
Explicitly sets the quantity of devices to renew for an offering, starting from the \code{effectiveDate} of the next period. The API returns a \code{NotEligible} error if the user is not permitted to invoke the operation. Please contact \href{mailto:aws-devicefarm-support@amazon.com}{aws-devicefarm-support@amazon.com} if you believe that you should be able to invoke this operation.
}
\examples{
# The following example renews a specific device slot offering.
renew_offering(
  offeringId = "D68B3C05-1BA6-4360-BC69-12345EXAMPLE",
  quantity = 1L
)

}
