% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{schedule_run}
\alias{schedule_run}
\title{Schedules a run}
\usage{
schedule_run(projectArn, appArn = NULL, devicePoolArn, name = NULL,
  test, configuration = NULL, executionConfiguration = NULL)
}
\arguments{
\item{projectArn}{[required] The ARN of the project for the run to be scheduled.}

\item{appArn}{The ARN of the app to schedule a run.}

\item{devicePoolArn}{[required] The ARN of the device pool for the run to be scheduled.}

\item{name}{The name for the run to be scheduled.}

\item{test}{[required] Information about the test for the run to be scheduled.}

\item{configuration}{Information about the settings for the run to be scheduled.}

\item{executionConfiguration}{Specifies configuration information about a test run, such as the execution timeout (in minutes).}
}
\description{
Schedules a run.
}
\section{Accepted Parameters}{
\preformatted{schedule_run(
  projectArn = "string",
  appArn = "string",
  devicePoolArn = "string",
  name = "string",
  test = list(
    type = "BUILTIN_FUZZ"|"BUILTIN_EXPLORER"|"WEB_PERFORMANCE_PROFILE"|"APPIUM_JAVA_JUNIT"|"APPIUM_JAVA_TESTNG"|"APPIUM_PYTHON"|"APPIUM_WEB_JAVA_JUNIT"|"APPIUM_WEB_JAVA_TESTNG"|"APPIUM_WEB_PYTHON"|"CALABASH"|"INSTRUMENTATION"|"UIAUTOMATION"|"UIAUTOMATOR"|"XCTEST"|"XCTEST_UI"|"REMOTE_ACCESS_RECORD"|"REMOTE_ACCESS_REPLAY",
    testPackageArn = "string",
    testSpecArn = "string",
    filter = "string",
    parameters = list(
      "string"
    )
  ),
  configuration = list(
    extraDataPackageArn = "string",
    networkProfileArn = "string",
    locale = "string",
    location = list(
      latitude = 123.0,
      longitude = 123.0
    ),
    vpceConfigurationArns = list(
      "string"
    ),
    customerArtifactPaths = list(
      iosPaths = list(
        "string"
      ),
      androidPaths = list(
        "string"
      ),
      deviceHostPaths = list(
        "string"
      )
    ),
    radios = list(
      wifi = TRUE|FALSE,
      bluetooth = TRUE|FALSE,
      nfc = TRUE|FALSE,
      gps = TRUE|FALSE
    ),
    auxiliaryApps = list(
      "string"
    ),
    billingMethod = "METERED"|"UNMETERED"
  ),
  executionConfiguration = list(
    jobTimeoutMinutes = 123,
    accountsCleanup = TRUE|FALSE,
    appPackagesCleanup = TRUE|FALSE,
    videoCapture = TRUE|FALSE,
    skipAppResign = TRUE|FALSE
  )
)
}
}

\examples{
# The following example schedules a test run named MyRun.
\donttest{schedule_run(
  name = "MyRun",
  devicePoolArn = "arn:aws:devicefarm:us-west-2:123456789101:pool:EXAMPLE-GUID-123-456",
  projectArn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  test = list(
    type = "APPIUM_JAVA_JUNIT",
    testPackageArn = "arn:aws:devicefarm:us-west-2:123456789101:test:EXAMPLE-GUID-123-456"
  )
)}

}
