% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{stop_remote_access_session}
\alias{stop_remote_access_session}
\title{Ends a specified remote access session}
\usage{
stop_remote_access_session(
  arn = "string"
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the remote access session you wish to stop.}
}
\description{
Ends a specified remote access session.
}
\examples{

}
