% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{stop_run}
\alias{stop_run}
\title{Initiates a stop request for the current test run}
\usage{
stop_run(arn)
}
\arguments{
\item{arn}{Represents the Amazon Resource Name (ARN) of the Device Farm run you wish to stop.}
}
\description{
Initiates a stop request for the current test run. AWS Device Farm will immediately stop the run on devices where tests have not started executing, and you will not be billed for these devices. On devices where tests have started executing, Setup Suite and Teardown Suite tests will run to completion before stopping execution on those devices. You will be billed for Setup, Teardown, and any tests that were in progress or already completed.
}
\examples{
# The following example stops a specific test run.
stop_run(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:run:EXAMPLE-GUID-123-456"
)

}
