% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{update_device_instance}
\alias{update_device_instance}
\title{Updates information about an existing private device instance}
\usage{
update_device_instance(arn, profileArn = NULL, labels = NULL)
}
\arguments{
\item{arn}{The Amazon Resource Name (ARN) of the device instance.}

\item{profileArn}{The Amazon Resource Name (ARN) of the profile that you want to associate with the device instance.}

\item{labels}{An array of strings that you want to associate with the device instance.}
}
\description{
Updates information about an existing private device instance.
}
\examples{

}
