% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{update_instance_profile}
\alias{update_instance_profile}
\title{Updates information about an existing private device instance profile}
\usage{
update_instance_profile(arn, name = NULL, description = NULL,
  packageCleanup = NULL, excludeAppPackagesFromCleanup = NULL,
  rebootAfterUse = NULL)
}
\arguments{
\item{arn}{The Amazon Resource Name (ARN) of the instance profile.}

\item{name}{The updated name for your instance profile.}

\item{description}{The updated description for your instance profile.}

\item{packageCleanup}{The updated choice for whether you want to specify package cleanup. The default value is \code{false} for private devices.}

\item{excludeAppPackagesFromCleanup}{An array of strings specifying the list of app packages that should not be cleaned up from the device after a test run is over.

The list of packages is only considered if you set \code{packageCleanup} to \code{true}.}

\item{rebootAfterUse}{The updated choice for whether you want to reboot the device after use. The default value is \code{true}.}
}
\description{
Updates information about an existing private device instance profile.
}
\examples{

}
