% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{update_project}
\alias{update_project}
\title{Modifies the specified project name, given the project ARN and a new name}
\usage{
update_project(
  arn = "string",
  name = "string",
  defaultJobTimeoutMinutes = 123
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the project whose name you wish to update.}

\item{name}{A string representing the new name of the project that you are updating.}

\item{defaultJobTimeoutMinutes}{The number of minutes a test run in the project will execute before it times out.}
}
\description{
Modifies the specified project name, given the project ARN and a new name.
}
\examples{
# The following example updates the specified project with a new name.
update_project(
  name = "NewName",
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:8f75187d-101e-4625-accc-12345EXAMPLE"
)

}
