% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{update_vpce_configuration}
\alias{update_vpce_configuration}
\title{Updates information about an existing Amazon Virtual Private Cloud (VPC) endpoint configuration}
\usage{
update_vpce_configuration(arn, vpceConfigurationName = NULL,
  vpceServiceName = NULL, serviceDnsName = NULL,
  vpceConfigurationDescription = NULL)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the VPC endpoint configuration you want to update.}

\item{vpceConfigurationName}{The friendly name you give to your VPC endpoint configuration, to manage your configurations more easily.}

\item{vpceServiceName}{The name of the VPC endpoint service running inside your AWS account that you want Device Farm to test.}

\item{serviceDnsName}{The DNS (domain) name used to connect to your private service in your Amazon VPC. The DNS name must not already be in use on the Internet.}

\item{vpceConfigurationDescription}{An optional description, providing more details about your VPC endpoint configuration.}
}
\description{
Updates information about an existing Amazon Virtual Private Cloud (VPC) endpoint configuration.
}
\section{Accepted Parameters}{
\preformatted{update_vpce_configuration(
  arn = "string",
  vpceConfigurationName = "string",
  vpceServiceName = "string",
  serviceDnsName = "string",
  vpceConfigurationDescription = "string"
)
}
}

