% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{delete_network_profile}
\alias{delete_network_profile}
\title{Deletes a network profile}
\usage{
delete_network_profile(arn)
}
\arguments{
\item{arn}{The Amazon Resource Name (ARN) of the network profile you want to delete.}
}
\description{
Deletes a network profile.
}
\examples{

}
