% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_account_settings}
\alias{get_account_settings}
\title{Returns the number of unmetered iOS and/or unmetered Android devices that have been purchased by the account}
\usage{
get_account_settings()
}
\description{
Returns the number of unmetered iOS and/or unmetered Android devices that have been purchased by the account.
}
\examples{
# The following example returns information about your Device Farm account
# settings.
get_account_settings()

}
