% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_job}
\alias{get_job}
\title{Gets information about a job}
\usage{
get_job(arn)
}
\arguments{
\item{arn}{The job's ARN.}
}
\description{
Gets information about a job.
}
\examples{
# The following example returns information about a specific job.
get_job(
  arn = "arn:aws:devicefarm:us-west-2::job:123-456-EXAMPLE-GUID"
)

}
