% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{get_test}
\alias{get_test}
\title{Gets information about a test}
\usage{
get_test(arn)
}
\arguments{
\item{arn}{The test's ARN.}
}
\description{
Gets information about a test.
}
\examples{
# The following example gets information about a specific test.
get_test(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:test:EXAMPLE-GUID-123-456"
)

}
