% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{list_devices}
\alias{list_devices}
\title{Gets information about unique device types}
\usage{
list_devices(arn = NULL, nextToken = NULL)
}
\arguments{
\item{arn}{The Amazon Resource Name (ARN) of the project.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}
}
\description{
Gets information about unique device types.
}
\examples{
# The following example returns information about the available devices in
# a specific project.
list_devices(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)

}
