% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{schedule_run}
\alias{schedule_run}
\title{Schedules a run}
\usage{
schedule_run(projectArn, devicePoolArn, test, appArn = NULL, name = NULL,
  configuration = NULL, executionConfiguration = NULL)
}
\arguments{
\item{projectArn}{The ARN of the project for the run to be scheduled.}

\item{devicePoolArn}{The ARN of the device pool for the run to be scheduled.}

\item{test}{Information about the test for the run to be scheduled.}

\item{appArn}{The ARN of the app to schedule a run.}

\item{name}{The name for the run to be scheduled.}

\item{configuration}{Information about the settings for the run to be scheduled.}

\item{executionConfiguration}{Specifies configuration information about a test run, such as the execution timeout (in minutes).}
}
\description{
Schedules a run.
}
\examples{
# The following example schedules a test run named MyRun.
schedule_run(
  name = "MyRun",
  devicePoolArn = "arn:aws:devicefarm:us-west-2:123456789101:pool:EXAMPLE-GUID-123-456",
  projectArn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456",
  test = list(
    type = "APPIUM_JAVA_JUNIT",
    testPackageArn = "arn:aws:devicefarm:us-west-2:123456789101:test:EXAMPLE-GUID-123-456"
  )
)

}
