% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.devicefarm_operations.R
\name{update_device_pool}
\alias{update_device_pool}
\title{Modifies the name, description, and rules in a device pool given the attributes and the pool ARN}
\usage{
update_device_pool(arn, name = NULL, description = NULL,
  rules = NULL)
}
\arguments{
\item{arn}{[required] The Amazon Resourc Name (ARN) of the Device Farm device pool you wish to update.}

\item{name}{A string representing the name of the device pool you wish to update.}

\item{description}{A description of the device pool you wish to update.}

\item{rules}{Represents the rules you wish to modify for the device pool. Updating rules is optional; however, if you choose to update rules for your request, the update will replace the existing rules.}
}
\description{
Modifies the name, description, and rules in a device pool given the attributes and the pool ARN. Rule updates are all-or-nothing, meaning they can only be updated as a whole (or not at all).
}
\section{Accepted Parameters}{
\preformatted{update_device_pool(
  arn = "string",
  name = "string",
  description = "string",
  rules = list(
    list(
      attribute = "ARN"|"PLATFORM"|"FORM_FACTOR"|"MANUFACTURER"|"REMOTE_ACCESS_ENABLED"|"REMOTE_DEBUG_ENABLED"|"APPIUM_VERSION"|"INSTANCE_ARN"|"INSTANCE_LABELS"|"FLEET_TYPE",
      operator = "EQUALS"|"LESS_THAN"|"GREATER_THAN"|"IN"|"NOT_IN"|"CONTAINS",
      value = "string"
    )
  )
)
}
}

\examples{
# The following example updates the specified device pool with a new name
# and description. It also enables remote access of devices in the device
# pool.
\donttest{update_device_pool(
  name = "NewName",
  arn = "arn:aws:devicefarm:us-west-2::devicepool:082d10e5-d7d7-48a5-ba5c-12345EXAMPLE",
  description = "NewDescription",
  rules = list(
    list(
      value = "True",
      attribute = "REMOTE_ACCESS_ENABLED",
      operator = "EQUALS"
    )
  )
)}

}
